api/requirements.txt
alembic==1.12.1
aiofiles==23.2.1
fastapi==0.104.1
fastapi_sqlalchemy==0.2.1
pydantic==2.4.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
uvicorn==0.24.0.post1
python-multipart==0.0.6
psycopg2==2.9.9